SPARKC-155: Repart uses only DC Local Hosts
This fixes a small bug where the repartition code could
select a replica for a give C* key that was not resident in the
same datacenter as the one the connector had chosen. This would
result in a random assignment of a node in the Datacenter. The new
code only chooses from replicas which are in the same datacenter
as the CassandraConnector.